requirements.txt
environs==5.0.0
PyGithub==1.43.7
